% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JCRImpactFactor.R
\name{JCRImpactFactor}
\alias{JCRImpactFactor}
\title{Journal Citation Reports (JCR) Impact Factor by Clarivate Analytics}
\description{
This package provides the Impact Factor of a journal reported by Journal
Citation Reports (JCR) of  Clarivate Analytics. The impact factor is
provided for only those journals that are included Journal
Citation Reports (JCR).
}
\details{
\tabular{ll}{ Package: JCRImpactFactor \tab JCR\cr Version: \tab 1.0.0 \cr Date: \tab
2021-04-20\cr Depends: \tab R (>= 3.5.0) \cr License: \tab GPL (>= 3) }
}
\references{
InCites journal citation reports, clarivate analytics.
}
\author{
Shahla Faisal  \href{mailto:shahla_ramzan@yahoo.com}{shahla_ramzan@yahoo.com}
}
\keyword{ImpactFactor}
\keyword{JCR}
\keyword{analytics}
\keyword{clarivate}
\keyword{package}
